Fix compile errors for v6.2.0 release build
- Fix KDoc bracket parsing in AdblockRuleParser
- Add DNS_PROXY branch to all exhaustive when(BlockMethod) blocks
- Fix WgConfig constructor params in DnsVpnService (String->ByteArray)
- Fix WireGuardProxy Long!=Int comparison
- Fix Vico 2.0 axis API (VerticalAxis.rememberStart/HorizontalAxis.rememberBottom)
- Fix Glance widget ColorProvider constructor (single Color, not day/night)
- Fix coroutine isActive ambiguity in HomeViewModel/StatsScreen
- Add missing imports (DnsProxyService, collectAsStateWithLifecycle)
- Fix suspend function calls in DnsProxyService/LocalDnsServer
- Fix ConnectionTracker synchronized return scope
- Add @OptIn for FlowRow ExperimentalLayoutApi
- Fix QrConfigScreen useDoH -> dohEnabled preference name
- Remove broken captivePortalApiUrl reference

+++ app/app/src/main/java/com/hostshield/service/WireGuardProxy.kt
@@ -137,7 +137,7 @@ class WireGuardProxy @Inject constructor() {
             }
 
             val msgType = readUint32LE(recvBuf, 0)
-            if (msgType != HANDSHAKE_RESPONSE) {
+            if (msgType != HANDSHAKE_RESPONSE.toLong()) {
                 Log.w(TAG, "Unexpected message type: $msgType")
                 return@withContext false
             }
